extracted SelectionGenerator to a new file
