+++ java-backend/src/main/java/org/kframework/backend/java/symbolic/SymbolicUnifier.java
@@ -671,7 +671,7 @@ public class SymbolicUnifier extends AbstractUnifier {
 
                 try {
                     for (int i = 0; i < otherCells.length; ++i) {
-                        unify(cells[generator.selection.get(i)], otherCells[i]);
+                        unify(cells[generator.getSelection(i)], otherCells[i]);
                     }
                 } catch (UnificationFailure e) {
                     continue;
@@ -679,7 +679,7 @@ public class SymbolicUnifier extends AbstractUnifier {
 
                 CellCollection.Builder builder = CellCollection.builder(context);
                 for (int i = 0; i < cells.length; ++i) {
-                    if (!generator.selected.contains(i)) {
+                    if (!generator.isSelected(i)) {
                         builder.add(cells[i]);
                     }
                 }
@@ -724,65 +724,6 @@ public class SymbolicUnifier extends AbstractUnifier {
         };
 
         return Multimaps.filterKeys(cellCollection.cellMap(), notRemoved);
-    }
-
-    private class SelectionGenerator {
-
-        private final int size;
-        private final int coSize;
-        public List<Integer> selection;
-        public Set<Integer> selected;
-        private int index;
-
-        public SelectionGenerator(int size, int coSize) {
-            assert size <= coSize;
-
-            this.size = size;
-            this.coSize = coSize;
-            selection = new ArrayList<Integer>();
-            selected = new HashSet<Integer>();
-            for (int i = 0; i < size; ++i) {
-                selection.add(i);
-                selected.add(i);
-            }
-        }
-
-        private void pop() {
-            index = selection.remove(selection.size() - 1);
-            selected.remove(index);
-            ++index;
-        }
-
-        private void push() {
-            selection.add(index);
-            selected.add(index);
-            index = 0;
-        }
-
-        public boolean generate() {
-            if (selection.isEmpty()) return false;
-            pop();
-            while (selection.size() != size) {
-                if (index == coSize) {
-                    if (selection.isEmpty()) {
-                        break;
-                    } else {
-                        pop();
-                        continue;
-                    }
-                }
-
-                if (!selected.contains(index)) {
-                    push();
-                    continue;
-                }
-
-                ++index;
-            }
-
-            return !selection.isEmpty();
-        }
-
     }
 
     @Override
